+++ db/migrations/0024_fearless_avengers.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "service_variant" ADD COLUMN "discount_price" numeric(12, 2);